Moved `checkin` command to reload `Checkin` Cog
